pyproject.toml
[project]
name = "vpn-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "greenlet (>=3.2.1,<4.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytz (>=2025.2,<2026.0)"
]
src_paths = ["src"]
virtual_env = "env"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.8"

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false

[tool.ruff]
line-length = 88
extend-select = ["I", "E", "F"]
